Failed attempt to control tree growth speed
	Added a Grwoth class with a NewPosition function to be used instead of Lerp to grow something over a fixed duration.
This should work in principle but my initial attempt faild because the growth rate parameter was not being correctly passed to objects created using coroutines.  I was probably just going about it the wrong way but am tempted to re-write all of the fractal code in a more functional style.